+++ a2/processing.h
@@ -10,9 +10,9 @@ int directScore(char* answer, char* likert_options[]);
 int reverseScore(char* answer, char* likert_options[]);
 
 void calculateScores(Respondent* responses[], int response_count, char* likert_options[]);
-float* calcPercentage(Respondent* responses[], int response_count, Question q);
+double* calcPercentage(Respondent* responses[], int response_count, Question q);
 
-void calcAverages(Respondent* responses[], int response_count, float output[], int output_size);
-void calcAveragesIndividual(Respondent* r, float output[], int output_size);
+void calcAverages(Respondent* responses[], int response_count, double output[], int output_size);
+void calcAveragesIndividual(Respondent* r, double output[], int output_size);
 
 #endif
